pyproject.toml
[project]
name = "phontex"
version = "0.1.0"
description = "IPA OCR desktop tool — screenshot to clipboard"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "peft>=0.10.0",
    "accelerate>=0.20.0",
    "pillow>=10.0.0",
    "editdistance>=0.6.0",
    "tqdm>=4.60.0",
]

[project.optional-dependencies]
dev = ["ruff", "pytest"]
frontend-server = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pyperclip>=1.9.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]

[tool.hatch.build.targets.wheel]
packages = ["ipa_ocr"]

[tool.ruff]
line-length = 100
target-version = "py310"